move utils.py functions to __init__.py
