+++ src/bits/bips/bip32.py
@@ -7,6 +7,7 @@ import hmac
 from typing import Tuple
 from typing import Union
 
+from bits import pubkey
 from bits.base58 import base58check
 from bits.base58 import base58check_decode
 from bits.ecmath import add_mod_p
@@ -17,7 +18,6 @@ from bits.ecmath import privkey_int
 from bits.ecmath import SECP256K1_Gx
 from bits.ecmath import SECP256K1_Gy
 from bits.ecmath import SECP256K1_N
-from bits.utils import pubkey
 
 VERSION_PUBLIC_MAINNET = b"\x04\x88\xb2\x1e"
 VERSION_PRIVATE_MAINNET = b"\x04\x88\xAD\xE4"
